+++ src/dve/core_engine/backends/implementations/spark/contract.py
@@ -7,7 +7,7 @@ from typing import Any, Optional
 from uuid import uuid4
 
 from pydantic import BaseModel
-from pydantic.fields import ModelField
+from pydantic.fields import FieldInfo
 from pyspark.sql import DataFrame, SparkSession
 from pyspark.sql import functions as sf
 from pyspark.sql.functions import col, lit
@@ -104,7 +104,9 @@ class SparkDataContract(BaseDataContract[DataFrame]):
 
         successful = True
         for entity_name, record_df in entities.items():
-            entity_fields: dict[str, ModelField] = contract_metadata.schemas[entity_name].__fields__
+            entity_fields: dict[str, FieldInfo] = contract_metadata.schemas[
+                entity_name
+            ].model_fields
             spark_schema = get_type_from_annotation(contract_metadata.schemas[entity_name])
             spark_schema.add(StructField(RECORD_INDEX_COLUMN_NAME, LongType()))
             if df_is_empty(record_df):
@@ -147,19 +149,17 @@ class SparkDataContract(BaseDataContract[DataFrame]):
                 self.logger.info(f"Data contract found {msg_count} issues in {entity_name}")
 
             try:
-                # TODO: will need to revisit in pydantic v2 bump as model field no longer available
                 record_df = record_df.select(
                     *[
                         (
                             get_spark_cast_statement_from_annotation(
-                                fld, mdl_field.annotation
+                                fld, fld_info.annotation
                             ).alias(fld)
                             if fld in record_df.columns
                             else lit(None).cast(
-                                get_type_from_annotation(mdl_field.annotation).alias(fld)
-                            )
+                                get_type_from_annotation(fld_info.annotation)).alias(fld)
                         )
-                        for fld, mdl_field in entity_fields.items()
+                        for fld, fld_info in entity_fields.items()
                     ],
                     col(RECORD_INDEX_COLUMN_NAME).cast(LongType()).alias(RECORD_INDEX_COLUMN_NAME),
                 )