+++ tests/test_core_engine/test_backends/test_implementations/test_duckdb/test_audit_ddb.py
@@ -344,11 +344,11 @@ def test_get_error_report_submissions(ddb_audit_manager_threaded: DDBAuditingMan
     expected = [
         SubmissionInfo(
             submission_id=sub_1.submission_id,
-            **{fld: val for fld, val in sub_1.dict().items() if fld != "submission_id"},
+            **{fld: val for fld, val in sub_1.model_dump().items() if fld != "submission_id"},
         ),
         SubmissionInfo(
             submission_id=sub_3.submission_id,
-            **{fld: val for fld, val in sub_3.dict().items() if fld != "submission_id"},
+            **{fld: val for fld, val in sub_3.model_dump().items() if fld != "submission_id"},
         ),
     ]
     assert len(processed) == 2
@@ -420,8 +420,8 @@ def test_get_submission_status(ddb_audit_manager_threaded: DDBAuditingManager):
             ]
         )
         aud.add_submission_statistics_records([
-            SubmissionStatisticsRecord(submission_id=sub_1.submission_id, record_count=5, number_record_rejections=2, number_warnings=3),
-            SubmissionStatisticsRecord(submission_id=sub_4.submission_id, record_count=20, number_record_rejections=0, number_warnings=1)
+            SubmissionStatisticsRecord(submission_id=sub_1.submission_id, record_count=5, number_submission_rejections=0, number_record_rejections=2, number_warnings=3),
+            SubmissionStatisticsRecord(submission_id=sub_4.submission_id, record_count=20, number_submission_rejections=0, number_record_rejections=0, number_warnings=1)
         ])
 
     sub_stats_1 = ddb_audit_manager_threaded.get_submission_status(sub_1.submission_id)